Changing Computers

+++ BomberManOnline/BomberManOnline/src/Level.h
@@ -41,7 +41,7 @@ private:
 	void onBombExplosion(sf::Vector2f position, int explosionSize);
 
 	eCollidableTile getCollidableTile(sf::Vector2f position) const;
-	void changeCollidableTile(sf::Vector2i position, eCollidableTile collidableTile);
+	void changeCollidableTile(sf::Vector2f position, eCollidableTile collidableTile);
 	void addExplosionObject(sf::Vector2f position);
 	void kickBombToPosition(sf::Vector2f bombPosition, sf::Vector2f kickToPosition);
 	PlayerClient* getPlayer(int ID);
